+++ Services/IPersonService.cs
@@ -5,6 +5,6 @@ public interface IPersonService {
 
     Task<List<Person>> getPeople();
     Task savePerson(Person person);
-    Task<Person> getPersonWithNationalID(int NationalID);
+    Task<Person> getPersonWithNationalID(ulong NationalID);
     Task RemoveAsync(string id);
 }
